+++ HISTORY.md
@@ -7,9 +7,9 @@ All notable changes to this project will be documented in this file.
 <!-- add unreleased items here -->
 
 * Fixed
-  * Don't emit purls for non-standard package names without reachability qualifiers ([#])
+  * Don't emit purls for non-standard package names without reachability qualifiers ([#192])
 
-[#]: 
+[#192]: https://github.com/CycloneDX/cyclonedx-esbuild/pull/192
 
 ## 1.4.3 - 2026-08-10